完赛代码(有bug，程序有时会卡死在32任务，还有一次程序跑飞)

+++ hw/pid.h
@@ -4,7 +4,7 @@
 
 // variable 变量
 typedef float PID_ParType;
-typedef int PID_ReVarType;
+typedef double PID_ReVarType;
 typedef double PID_InVarType;
 
 struct _Structure_PIDCtrl{
